Changed to accept decimals

+++ src/Java_Level1/mathOperations.java
@@ -3,6 +3,7 @@ package Java_Level1;
 import java.util.Scanner;
 
 // Create a Method where the user can insert to numbers and make the 4 basic mathematical operations:
+// The Method should accept decimals
 
 public class mathOperations {
 
@@ -10,24 +11,24 @@ public class mathOperations {
         Scanner scanner = new Scanner(System.in);  // Create Scanner
        
         System.out.print("Please enter the first number: ");
-        int number1 = scanner.nextInt();           
+        double number1 = scanner.nextDouble();           
         System.out.println("Your first number is: " + number1);
 
         System.out.print("Please enter the Second number: ");
-        int number2 = scanner.nextInt();     
+        double number2 = scanner.nextDouble();     
         System.out.println("Your second number is: " + number2);
         scanner.close();  // Always close the scanner once user is done
 
-        int sum = number1 + number2;
+        double sum = number1 + number2;
         System.out.println("Sum: " + sum);
 
-        int subtract = number1 - number2;
+        double subtract = number1 - number2;
         System.out.println("Subtraction: " + subtract);
        
-        int division = number1 / number2;
+        double division = number1 / number2;
         System.out.println("Division: " + division);
 
-        int multiply = number1 * number2;
+        double multiply = number1 * number2;
         System.out.println("Multiplication: " + multiply);
     }       
 }
